Backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.1,<2.0
paho-mqtt>=2.1,<3.0
openpyxl>=3.1,<4.0
reportlab>=4.2,<5.0
djangorestframework-simplejwt>=5.4,<6.0
whitenoise>=6.8,<7.0
gunicorn>=23.0,<24.0
dj-database-url>=2.3,<3.0
